Implement rest layer for headway module

+++ src/main/java/life/coachy/backend/headway/domain/HeadwayService.java
@@ -1,6 +1,7 @@
 package life.coachy.backend.headway.domain;
 
 import java.util.Set;
+import life.coachy.backend.headway.domain.exception.HeadwayNotFoundException;
 import life.coachy.backend.headway.query.HeadwayQueryDto;
 import life.coachy.backend.headway.query.HeadwayQueryRepository;
 import org.bson.types.ObjectId;
@@ -19,11 +20,15 @@ class HeadwayService {
     this.headwayQueryRepository = headwayQueryRepository;
   }
 
-  void save(Headway headway) {
-    this.headwayRepository.save(headway);
+  Headway save(Headway headway) {
+    return this.headwayRepository.save(headway);
   }
 
   void deleteById(ObjectId id) {
+    if (!this.headwayQueryRepository.existsByIdentifier(id)) {
+      throw new HeadwayNotFoundException();
+    }
+
     this.headwayRepository.deleteById(id);
   }
